Remodela estrutura de dados
