+++ modulo3/labebank/src/Tipos.ts
@@ -1,0 +1,21 @@
+
+type Transaction = {
+            clientId:number,
+            value:number,
+            date:Date,
+            description:string
+    
+ }
+
+
+export type Count = {
+    id:number,
+    client: {
+        name:string,
+        cpf:string,
+        birthDate:string,
+        balance:number
+    },
+    extract: Transaction[]
+ 
+}
