--- modulo3/labebank/src/newClient.ts
@@ -1,24 +1,0 @@
-import { Client } from "./clientsList";
-
-export type NewCount = {
-    id:number,
-    client: {
-        id:number,
-        name:string,
-        cpf:string,
-        birthDate:string,
-        balance:number
-    },
-    extract: {
-        clientId:number,
-        transaction: { 
-            type:string,
-            value:number,
-            date:Date,
-            description:string
-    }
-    }
-}
-//tabela cliente --- 1 saldo
-//tabela transacao --- id cliente
-//
